refactor(lib): keep the shared test shim out of the published dist
`src/test/riteway.ts` sat next to `src/test/setup.ts`, which the library build
compiles — so a helper whose whole body imports `vitest`, a devDependency,
was being emitted into `dist`. Nothing imports it there, but shipping it invites
exactly one bad day.

`tsconfig.build.json` already excludes every `__tests__` directory, so the shim
moves into one. `dist/test` is back to `setup.js` alone.
